+++ viewer/include/icirv/report_model.hpp
@@ -5,7 +5,7 @@
 #include <string>
 #include <vector>
 
-#include "json_value.hpp"
+#include "icirv/json_value.hpp"
 
 namespace icirv {